feat: Created dynamically changing input field to change date
Added state to the input field and submit button so that when a user
types in a value and presses submit, the date that is in the title
of the app will change dynamically to the value in the input field.
This date will later determine the value of the countdown clock.